Rename maxBatchSize into maxBatchLogEvents
Amazon uses "maximum batch size" for measuring the size of a batch in
bytes. We used it for the number of events in the log. This commit
changes the name to be more consistent with Amazon's documentation.

See also: #8
See also: http://docs.aws.amazon.com/AmazonCloudWatchLogs/latest/APIReference/API_PutLogEvents.html